convert: Move data definition to dedicated 'export' module.
