+++ openage/convert/blendomatic.py
@@ -12,8 +12,8 @@ from struct import Struct, unpack_from
 
 from ..log import dbg
 from .dataformat.genie_structure import GenieStructure
-from .dataformat.data_definition import DataDefinition
-from .dataformat.struct_definition import StructDefinition
+from .export.data_definition import DataDefinition
+from .export.struct_definition import StructDefinition
 
 
 class BlendingTile:
