+++ openage/convert/export/CMakeLists.txt
@@ -1,3 +1,12 @@
 add_py_modules(
 	__init__.py
+	content_snippet.py
+	data_definition.py
+	data_formatter.py
+	entry_parser.py
+	generated_file.py
+	header_snippet.py
+	struct_definition.py
+	struct_snippet.py
+	util.py
 )
