+++ openage/convert/export/struct_snippet.py
